requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
torch==2.1.0
torchvision==0.16.0
facenet-pytorch==2.5.3
Pillow==10.1.0
numpy==1.26.2
PyJWT==2.8.0